refactor(structure): refactored routes and some absolute imports, renamed routes folder to pages
