+++ src/pages/kuski/PlayedBattles.js
@@ -2,13 +2,15 @@ import React from 'react';
 import { graphql, compose } from 'react-apollo';
 import TablePagination from '@material-ui/core/TablePagination';
 import PropTypes from 'prop-types';
-import { sortResults } from 'utils';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
-import Time from '../../components/Time';
-import Link from '../../components/Link';
-import Kuski from '../../components/Kuski';
-import LocalTime from '../../components/LocalTime';
-import { BattleType } from '../../components/Names';
+
+import Time from 'components/Time';
+import Link from 'components/Link';
+import Kuski from 'components/Kuski';
+import LocalTime from 'components/LocalTime';
+import { BattleType } from 'components/Names';
+import { sortResults } from 'utils/battle';
+
 import battlesQuery from './battles.graphql';
 import s from './PlayedBattles.css';
