+++ src/utils/upload.js
@@ -4,8 +4,8 @@ import fs from 'fs';
 import crypto from 'crypto';
 import generate from 'nanoid/generate';
 import AWS from 'aws-sdk';
-import { Level, Replay as ReplayDB } from './data/models';
-import config from './config';
+import { Level, Replay as ReplayDB } from '../data/models';
+import config from '../config';
 
 const getLevelsFromName = async LevelName => {
   const levels = await Level.findAll({
